Making a calculator to find the value of an option
In this file, we will write code to extract option contracts from a dataframe, calculate the value of each option, and then insert them all into a full new column of the dataset and export it to a separate file. We will also plot heatmaps to visualise how the Strike Price and Volatility affect the value of an option.